Move Paginator class to pagination.py
