+++ src/commands/viewbinds.py
@@ -3,13 +3,13 @@ from resources.bloxlink import instance as bloxlink
 from resources.groups import get_group
 from resources.models import CommandContext
 from resources.constants import RED_COLOR
-from resources.pagination import button_author_validation
+from resources.pagination import button_author_validation, Paginator
 from resources.component_helper import get_custom_id_data, set_components
 from resources.exceptions import RobloxAPIError
 import hikari
 
 
-MAX_BINDS_PER_PAGE = 2
+MAX_BINDS_PER_PAGE = 3
 
 
 async def viewbinds_category_autocomplete(interaction: hikari.AutocompleteInteraction):
@@ -206,77 +206,6 @@ def viewbinds_item_filter(id_filter, category_filter):
         return json_binds_to_guild_binds(items, category=category_filter, id_filter=id_filter)
 
     return wrapper
-
-
-class Paginator:
-    def __init__(
-        self,
-        guild_id,
-        user_id,
-        items,
-        page_number=0,
-        max_items=10,
-        custom_formatter=None,
-        extra_custom_ids="",
-        item_filter=None,
-    ):
-        self.guild_id = guild_id
-        self.user_id = user_id
-        self.page_number = page_number
-        self.items = items if not item_filter else item_filter(items)
-        self.max_items = max_items
-        self.custom_formatter = custom_formatter
-        self.extra_custom_ids = extra_custom_ids
-
-    @property
-    async def embed(self):
-        offset = self.page_number * self.max_items
-        max_items = (
-            len(self.items) if (offset + self.max_items >= len(self.items)) else offset + self.max_items
-        )
-        current_items = self.items[offset:max_items]
-
-        if self.custom_formatter:
-            embed = await self.custom_formatter(self.page_number, current_items, self.guild_id)
-        else:
-            embed = hikari.Embed(title=f"Test Pagination", description=f"Page {self.page_number}")
-
-        return embed
-
-    @embed.setter
-    def embed(self, value):
-        self._embed = value
-
-    @property
-    def components(self):
-        button_row = bloxlink.rest.build_message_action_row()
-
-        offset = self.page_number * self.max_items
-        max_items = (
-            len(self.items) if (offset + self.max_items >= len(self.items)) else offset + self.max_items
-        )
-
-        # Previous button
-        button_row.add_interactive_button(
-            hikari.ButtonStyle.SECONDARY,
-            f"viewbinds:{self.user_id}:{self.page_number-1}:{self.extra_custom_ids}",
-            label="\u2B9C",
-            is_disabled=True if self.page_number == 0 else False,
-        )
-
-        # Next button
-        button_row.add_interactive_button(
-            hikari.ButtonStyle.SECONDARY,
-            f"viewbinds:{self.user_id}:{self.page_number+1}:{self.extra_custom_ids}",
-            label="\u2B9E",
-            is_disabled=True if max_items == len(self.items) else False,
-        )
-
-        return button_row
-
-    @components.setter
-    def components(self, value):
-        self._components = value
 
 
 async def build_page_embed(page_components) -> hikari.Embed:
